refactor(pipeline): delete 4 legacy files + migrate hidden callers
Task 2.5: Delete classifier.py, doc_classifier.py (thin wrappers),
ingestion_worker.py (replaced by unified_pipeline), and
prepare_pdfs_full.py (1786 LOC monolith — all logic ported to modular
components in Phase 1 + Phase 2).

Migrate 3 hidden callers (document_scanner.py, monitoring.py,
documents.py) from wrapper imports to direct document_classifier imports.
Update test_upload_architecture.py to test unified_pipeline._emit.
Update classification/__init__.py to export hybrid_classifier.classify.
