--- backend/ai_engine/classification/doc_classifier.py
@@ -1,12 +1,0 @@
-"""Doc classifier — DELEGATION WRAPPER.
-
-REFACTOR (Phase 2, Step 7): All logic consolidated into
-ai_engine.classification.document_classifier.  This module re-exports
-for backward compatibility.
-"""
-from __future__ import annotations
-
-from ai_engine.classification.document_classifier import (  # noqa: F401
-    DOC_TYPES,
-    classify_registered_documents,
-)
